backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary>=2.9.10
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3
numpy==1.24.4
python-docx==1.1.0
reportlab==4.0.7

# NLP and ML
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.1

# Vector Database
pinecone-client==2.2.4
qdrant-client==1.7.0

# AWS S3
boto3==1.34.0
botocore==1.34.0

# Utilities
requests==2.31.0
httpx==0.25.2
pandas==2.1.4
python-dateutil==2.8.2

# Google OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# OpenAI API
openai==1.3.0

# PDF Processing
pdfplumber==0.10.3
